数解析问题

+++ DispatchSystem/DisplayForm.cs
@@ -89,7 +89,7 @@ namespace DispatchSystem
 
         private void button_0_Click(object sender, EventArgs e)
         {
-            UInt32 data = UdpSever.Ddata[deviceID, registerID, 0];
+            Int64 data = UdpSever.Ddata[deviceID, registerID, 0];
             data ^= 1;
             UdpSever.writeWord(deviceID, registerID, (UInt16)data);
             UdpSever.Ddata[deviceID, registerID, 0] = data;
@@ -97,7 +97,7 @@ namespace DispatchSystem
 
         private void button_1_Click(object sender, EventArgs e)
         {
-            UInt32 data = UdpSever.Ddata[deviceID, registerID, 0];
+            Int64 data = UdpSever.Ddata[deviceID, registerID, 0];
             data ^= 1 << 1;
             UdpSever.writeWord(deviceID, registerID, (UInt16)data);
             UdpSever.Ddata[deviceID, registerID, 0] = data;
@@ -105,7 +105,7 @@ namespace DispatchSystem
 
         private void button_2_Click(object sender, EventArgs e)
         {
-            UInt32 data = UdpSever.Ddata[deviceID, registerID, 0];
+            Int64 data = UdpSever.Ddata[deviceID, registerID, 0];
             data ^= 1 << 2;
             UdpSever.writeWord(deviceID, registerID, (UInt16)data);
             UdpSever.Ddata[deviceID, registerID, 0] = data;
@@ -113,7 +113,7 @@ namespace DispatchSystem
 
         private void button_3_Click(object sender, EventArgs e)
         {
-            UInt32 data = UdpSever.Ddata[deviceID, registerID, 0];
+            Int64 data = UdpSever.Ddata[deviceID, registerID, 0];
             data ^= 1 << 3;
             UdpSever.writeWord(deviceID, registerID, (UInt16)data);
             UdpSever.Ddata[deviceID, registerID, 0] = data;
@@ -121,7 +121,7 @@ namespace DispatchSystem
 
         private void button_4_Click(object sender, EventArgs e)
         {
-            UInt32 data = UdpSever.Ddata[deviceID, registerID, 0];
+            Int64 data = UdpSever.Ddata[deviceID, registerID, 0];
             data ^= 1 << 4;
             UdpSever.writeWord(deviceID, registerID, (UInt16)data);
             UdpSever.Ddata[deviceID, registerID, 0] = data;
@@ -129,7 +129,7 @@ namespace DispatchSystem
 
         private void button_5_Click(object sender, EventArgs e)
         {
-            UInt32 data = UdpSever.Ddata[deviceID, registerID, 0];
+            Int64 data = UdpSever.Ddata[deviceID, registerID, 0];
             data ^= 1 << 5;
             UdpSever.writeWord(deviceID, registerID, (UInt16)data);
             UdpSever.Ddata[deviceID, registerID, 0] = data;
@@ -137,7 +137,7 @@ namespace DispatchSystem
 
         private void button_6_Click(object sender, EventArgs e)
         {
-            UInt32 data = UdpSever.Ddata[deviceID, registerID, 0];
+            Int64 data = UdpSever.Ddata[deviceID, registerID, 0];
             data ^= 1 << 6;
             UdpSever.writeWord(deviceID, registerID, (UInt16)data);
             UdpSever.Ddata[deviceID, registerID, 0] = data;
@@ -145,7 +145,7 @@ namespace DispatchSystem
 
         private void button_7_Click(object sender, EventArgs e)
         {
-            UInt32 data = UdpSever.Ddata[deviceID, registerID, 0];
+            Int64 data = UdpSever.Ddata[deviceID, registerID, 0];
             data ^= 1 << 7;
             UdpSever.writeWord(deviceID, registerID, (UInt16)data);
             UdpSever.Ddata[deviceID, registerID, 0] = data;
@@ -153,7 +153,7 @@ namespace DispatchSystem
 
         private void button_8_Click(object sender, EventArgs e)
         {
-            UInt32 data = UdpSever.Ddata[deviceID, registerID, 0];
+            Int64 data = UdpSever.Ddata[deviceID, registerID, 0];
             data ^= 1 << 8;
             UdpSever.writeWord(deviceID, registerID, (UInt16)data);
             UdpSever.Ddata[deviceID, registerID, 0] = data;
@@ -161,7 +161,7 @@ namespace DispatchSystem
 
         private void button_9_Click(object sender, EventArgs e)
         {
-            UInt32 data = UdpSever.Ddata[deviceID, registerID, 0];
+            Int64 data = UdpSever.Ddata[deviceID, registerID, 0];
             data ^= 1 << 9;
             UdpSever.writeWord(deviceID, registerID, (UInt16)data);
             UdpSever.Ddata[deviceID, registerID, 0] = data;
@@ -169,7 +169,7 @@ namespace DispatchSystem
 
         private void button_10_Click(object sender, EventArgs e)
         {
-            UInt32 data = UdpSever.Ddata[deviceID, registerID, 0];
+            Int64 data = UdpSever.Ddata[deviceID, registerID, 0];
             data ^= 1 << 10;
             UdpSever.writeWord(deviceID, registerID, (UInt16)data);
             UdpSever.Ddata[deviceID, registerID, 0] = data;
@@ -177,7 +177,7 @@ namespace DispatchSystem
 
         private void button_11_Click(object sender, EventArgs e)
         {
-            UInt32 data = UdpSever.Ddata[deviceID, registerID, 0];
+            Int64 data = UdpSever.Ddata[deviceID, registerID, 0];
             data ^= 1 << 11;
             UdpSever.writeWord(deviceID, registerID, (UInt16)data);
             UdpSever.Ddata[deviceID, registerID, 0] = data;
@@ -185,7 +185,7 @@ namespace DispatchSystem
 
         private void button_12_Click(object sender, EventArgs e)
         {
-            UInt32 data = UdpSever.Ddata[deviceID, registerID, 0];
+            Int64 data = UdpSever.Ddata[deviceID, registerID, 0];
             data ^= 1 << 12;
             UdpSever.writeWord(deviceID, registerID, (UInt16)data);
             UdpSever.Ddata[deviceID, registerID, 0] = data;
@@ -193,7 +193,7 @@ namespace DispatchSystem
 
         private void button_13_Click(object sender, EventArgs e)
         {
-            UInt32 data = UdpSever.Ddata[deviceID, registerID, 0];
+            Int64 data = UdpSever.Ddata[deviceID, registerID, 0];
             data ^= 1 << 13;
             UdpSever.writeWord(deviceID, registerID, (UInt16)data);
             UdpSever.Ddata[deviceID, registerID, 0] = data;
@@ -201,7 +201,7 @@ namespace DispatchSystem
 
         private void button_14_Click(object sender, EventArgs e)
         {
-            UInt32 data = UdpSever.Ddata[deviceID, registerID, 0];
+            Int64 data = UdpSever.Ddata[deviceID, registerID, 0];
             data ^= 1 << 14;
             UdpSever.writeWord(deviceID, registerID, (UInt16)data);
             UdpSever.Ddata[deviceID, registerID, 0] = data;
@@ -209,7 +209,7 @@ namespace DispatchSystem
 
         private void button_15_Click(object sender, EventArgs e)
         {
-            UInt32 data = UdpSever.Ddata[deviceID, registerID, 0];
+            Int64 data = UdpSever.Ddata[deviceID, registerID, 0];
             data ^= 1 << 15;
             UdpSever.writeWord(deviceID, registerID, (UInt16)data);
             UdpSever.Ddata[deviceID, registerID, 0] = data;
